// apps/storage_management/src/types/enums/FileTypeEnum.ts
export enum FileTypeEnum {
  IMAGE = "image/jpeg",
  PNG = "image/png",
  VIDEO = "video/mp4",
  MP3 = "audio/mpeg",
  PDF = "application/pdf",
  TEXT = "text/plain",
  ZIP = "application/zip",
  CSV = "text/csv",
  EXCEL = "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
  WORD = "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
  POWERPOINT = "application/vnd.openxmlformats-officedocument.presentationml.presentation",
  JSON = "application/json",
  XML = "application/xml",
  HTML = "text/html",
  MARKDOWN = "text/markdown",
  SVG = "image/svg+xml",
  JS = "text/javascript",
  WAV = "audio/wav"
}